B0.4.x

+++ src/net/minecraft/client/shader/ShaderUniform.java
@@ -177,7 +177,6 @@ public class ShaderUniform extends ShaderDefault implements AutoCloseable {
 
    public void upload() {
       if (!this.dirty) {
-         ;
       }
 
       this.dirty = false;
@@ -187,7 +186,7 @@ public class ShaderUniform extends ShaderDefault implements AutoCloseable {
          this.uploadFloat();
       } else {
          if (this.uniformType > 10) {
-            LOGGER.warn("Uniform.upload called, but type value ({}) is not a valid type. Ignoring.", (int)this.uniformType);
+            LOGGER.warn("Uniform.upload called, but type value ({}) is not a valid type. Ignoring.", this.uniformType);
             return;
          }
 
@@ -212,7 +211,8 @@ public class ShaderUniform extends ShaderDefault implements AutoCloseable {
          OpenGlHelper.glUniform4iv(this.uniformLocation, this.uniformIntBuffer);
          break;
       default:
-         LOGGER.warn("Uniform.upload called, but count value ({}) is  not in the range of 1 to 4. Ignoring.", (int)this.uniformCount);
+         LOGGER.warn("Uniform.upload called, but count value ({}) is  not in the range of 1 to 4. Ignoring.",
+                 this.uniformCount);
       }
 
    }
@@ -233,7 +233,8 @@ public class ShaderUniform extends ShaderDefault implements AutoCloseable {
          OpenGlHelper.glUniform4fv(this.uniformLocation, this.uniformFloatBuffer);
          break;
       default:
-         LOGGER.warn("Uniform.upload called, but count value ({}) is not in the range of 1 to 4. Ignoring.", (int)this.uniformCount);
+         LOGGER.warn("Uniform.upload called, but count value ({}) is not in the range of 1 to 4. Ignoring.",
+                 this.uniformCount);
       }
 
    }
